+++ utils.py
@@ -1,0 +1,19 @@
+import os
+from base64 import b64encode
+from dotenv import load_dotenv
+
+
+# load environment variables from .env file
+load_dotenv()
+
+CRISP_IDENTIFIER = os.environ["CRISP_IDENTIFIER"]
+CRISP_KEY = os.environ["CRISP_KEY"]
+
+
+def get_auth_headers():
+    auth_string = f"{CRISP_IDENTIFIER}:{CRISP_KEY}"
+    auth_string_b64 = b64encode(auth_string.encode("utf-8")).decode("utf-8")
+    return {
+        "Authorization": f"Basic {auth_string_b64}",
+        "X-Crisp-Tier": "plugin",
+    }
